restructured Scenario API fns
